Move V1 types to V1 folder
